requirements.txt
Flask
python-dotenv
groq
requests